+++ packages/apollos-data-connector-postgres/src/postgres/migrations/index.js
@@ -1,0 +1,6 @@
+import AddUUID from './000_add_uuid_extension';
+import AddPGTRGM from './001_add_pg_trgm_extension';
+
+const migrations = [AddUUID, AddPGTRGM];
+
+export default migrations;