+++ packages/apollos-data-connector-postgres/src/user-likes/migrations/index.js
@@ -1,0 +1,5 @@
+import CreateUserLikes001 from './001_create_user_likes';
+
+const migrations = [CreateUserLikes001];
+
+export default migrations;